implemented final features

+++ src/api/deleteCreatorAPI.js
@@ -1,17 +1,20 @@
 import supabase from "../client";
 
-const deleteCreator = async (creatorName) => {
-  const { error } = await supabase
-    .from("creators")
-    .delete()
-    .eq("name", creatorName);
+const deleteCreator = async (req) => {
+  const { error } = await supabase.from("creators").delete().eq("url", req.url);
 
   if (error) {
-    console.error("Error deleting creator: ", error);
-    return false;
-  } else {
-    return true;
+    const res = {
+      task: false,
+      message: "Error is deleting creator, please try again!",
+    };
+    return res;
   }
+  const res = {
+    task: true,
+    message: "Creator deleted successfully",
+  };
+  return res;
 };
 
 export default deleteCreator;
